rag/ingestion/requirements.txt
boto3>=1.34.0
openai>=1.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
python-dotenv>=1.0.0
PyPDF2>=3.0.0  # For PDF support